[+] getSize

+++ include/Text.hpp
@@ -25,19 +25,19 @@ namespace Arcade {
             ObjectType getType() override;
             std::string & getText() override;
             std::string & getFont() override;
-            int getSize() override;
+            float getSize() override;
             color_t getColor() override;
             void setPos(pos_t pos);
             void setDisplay(bool display);
             void setText(const std::string &text);
             void setFont(const std::string &font);
-            void setSize(int size);
+            void setSize(float size);
             void setColor(color_t color);
 
         protected:
             std::string _font;
             std::string _text;
-            int _size;
+            float _size;
             color_t _color;
             pos_t _pos;
             bool _display;
